refactor and fix a small issue
